# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Movies Tracker API
  description: Comprehensive REST API for tracking movie watch history, managing catalogs, and annual leaderboards.
  version: 1.0.0
  contact:
    name: Siva B
    url: https://github.com/sivab193/movies-tracker
servers:
  - url: https://mt.siv19.dev/api
    description: Production Vercel Deployment
  - url: http://localhost:8000/api
    description: Local Development Server
components:
  securitySchemes:
    firebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID Token obtained after user authentication via Google OAuth.
  schemas:
...
    Movie:
      type: object
      properties:
        _id:
          type: string
        imdbId:
          type: string
        title:
          type: string
        year:
          type: string
        runtime:
          type: string
        runtimeSeconds:
          type: integer
        posterUrl:
          type: string
        genre:
          type: string
        director:
          type: string
        imdbRating:
          type: string
...
paths:
...
  /movies:
    get:
      summary: Get paginated list of movies
      parameters:
        - in: query
          name: skip
          schema:
            type: integer
            default: 0
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of movies and total count
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  movies:
                    type: array
                    items:
                      $ref: '#/components/schemas/Movie'
    post:
      summary: Add movie from OMDb (Admin only)
      security:
        - firebaseAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - imdbId
              properties:
                imdbId:
                  type: string
                  example: tt15398776
      responses:
        '201':
          description: Movie added successfully
  /movies/{movieId}/poster:
    get:
      summary: Get binary poster image stored in MongoDB
      parameters:
        - in: path
          name: movieId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Binary image stream
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
...